+++ jpsonic-main/src/main/java/org/airsonic/player/service/search/AnalyzerFactory.java
@@ -1,24 +1,25 @@
 /*
- This file is part of Jpsonic.
+ This file is part of Airsonic.
 
- Jpsonic is free software: you can redistribute it and/or modify
+ Airsonic is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
  the Free Software Foundation, either version 3 of the License, or
  (at your option) any later version.
 
- Jpsonic is distributed in the hope that it will be useful,
+ Airsonic is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  GNU General Public License for more details.
 
  You should have received a copy of the GNU General Public License
- along with Jpsonic.  If not, see <http://www.gnu.org/licenses/>.
+ along with Airsonic.  If not, see <http://www.gnu.org/licenses/>.
 
- Copyright 2019 (C) tesshu.com
+ Copyright 2016 (C) Airsonic Authors
+ Based upon Subsonic, Copyright 2009 (C) Sindre Mehus
  */
-package com.tesshu.jpsonic.service.search;
 
-import com.tesshu.jpsonic.service.search.IndexType.FieldNames;
+package org.airsonic.player.service.search;
+
 import com.tesshu.jpsonic.service.search.analysis.GenreTokenizerFactory;
 import com.tesshu.jpsonic.service.search.analysis.HiraganaStopFilterFactory;
 import com.tesshu.jpsonic.service.search.analysis.Id3ArtistTokenizerFactory;
@@ -39,43 +40,24 @@ import org.apache.lucene.analysis.miscellaneous.PerFieldAnalyzerWrapper;
 import org.apache.lucene.analysis.pattern.PatternReplaceFilterFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Analyzer provider. This class is a division of what was once part of
- * SearchService and added functionality.
- * 
- * Analyzer can be closed but is a reuse premise. It is held in this class.
- * 
- * Some of the query parses performed in legacy services are defined in
- * QueryAnalyzer. Most of the actual query parsing is done with QueryFactory.
+ * Analyzer provider.
+ * This class is a division of what was once part of SearchService and added functionality.
+ * This class provides Analyzer which is used at index generation
+ * and QueryAnalyzer which analyzes the specified query at search time.
+ * Analyzer can be closed but is a reuse premise.
+ * It is held in this class.
  */
+@Component
 public final class AnalyzerFactory {
 
-    private static AnalyzerFactory instance;
-
     private static final Logger LOG = LoggerFactory.getLogger(AnalyzerFactory.class);
-    
-    private static final Object l = new Object();
-
-    /**
-     * Returns an instance of AnalyzerFactory.
-     * 
-     * @return AnalyzerFactory instance
-     */
-    public static AnalyzerFactory getInstance() {
-        if (null == instance) {
-            synchronized (l) {
-                if (instance == null) {
-                    instance = new AnalyzerFactory();
-                }
-            }
-        }
-        return instance;
-    }
 
     private Analyzer analyzer;
 
@@ -84,9 +66,6 @@ public final class AnalyzerFactory {
     public static final String STOP_TAGS = "org/apache/lucene/analysis/ja/stoptags.txt";
 
     public static final String STOP_WARDS = "com/tesshu/jpsonic/service/stopwords.txt";
-
-    private AnalyzerFactory() {
-    }
 
     private CustomAnalyzer.Builder addWildCard(CustomAnalyzer.Builder builder) throws IOException {
         builder.addTokenFilter(PatternReplaceFilterFactory.class, "pattern", "(^(?!.*\\*$).+$)", "replacement", "$1*", "replace", "first");