+++ jpsonic-main/src/main/java/org/airsonic/player/validator/UserSettingsValidator.java
@@ -24,23 +24,27 @@ import org.airsonic.player.controller.UserSettingsController;
 import org.airsonic.player.service.SecurityService;
 import org.airsonic.player.service.SettingsService;
 import org.apache.commons.lang.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * Validator for {@link UserSettingsController}.
  *
  * @author Sindre Mehus
  */
-@Component
 public class UserSettingsValidator implements Validator {
 
-    @Autowired
     private SecurityService securityService;
-    @Autowired
     private SettingsService settingsService;
+    private HttpServletRequest request;
+
+    public UserSettingsValidator(SecurityService securityService, SettingsService settingsService, HttpServletRequest request) {
+        this.securityService = securityService;
+        this.settingsService = settingsService;
+        this.request = request;
+    }
 
     /**
      * {@inheritDoc}
@@ -83,6 +87,16 @@ public class UserSettingsValidator implements Validator {
 
         if (command.isPasswordChange() && command.isLdapAuthenticated()) {
             errors.rejectValue("password", "usersettings.passwordnotsupportedforldap");
+        }
+
+        if (securityService.getCurrentUser(request).getUsername().equals(username)) {
+            // These errors don't need translation since the option isn't exposed to the user
+            if (command.isDeleteUser()) {
+                errors.rejectValue("deleteUser", null, "Cannot delete the current user");
+            }
+            if (! command.isAdminRole()) {
+                errors.rejectValue("adminRole", null, "Cannot remove admin from the current user");
+            }
         }
 
     }